requirements.txt
joblib==1.2.0
numpy==1.24.3
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
